requirements.txt
causal_conv1d==1.1.1
easydict==1.13
einops==0.8.1
h5py==3.11.0
KNN_CUDA==0.2
mamba_ssm==1.1.1
numpy==1.24.3
PyYAML==6.0.2
scipy==1.9.3
timm==1.0.19
torch_geometric==2.6.1
torch_scatter==2.1.2
tqdm==4.66.5
transformers==4.45.1
